extensions: Always compile examples to prevent them from breaking
