--- osquery/examples/example_test.cpp
@@ -1,24 +1,0 @@
-/**
- * Copyright (c) 2014-present, The osquery authors
- *
- * This source code is licensed as defined by the LICENSE file found in the
- * root directory of this source tree.
- *
- * SPDX-License-Identifier: (Apache-2.0 OR GPL-2.0-only)
- */
-
-#include <gtest/gtest.h>
-
-namespace osquery {
-namespace example {
-
-class ExampleTests : public testing::Test {};
-
-TEST_F(ExampleTests, test_plugin) { EXPECT_TRUE(1 == 1); }
-}
-}
-
-int main(int argc, char* argv[]) {
-  testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
