--- osquery/examples/extension_group_example/src/example.h
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2014-present, The osquery authors
- *
- * This source code is licensed as defined by the LICENSE file found in the
- * root directory of this source tree.
- *
- * SPDX-License-Identifier: (Apache-2.0 OR GPL-2.0-only)
- */
-
-#pragma once
-
-#include <osquery/core/system.h>
-#include <osquery/sdk.h>
-
-namespace osquery {
-class ExampleTable : public TablePlugin {
- private:
-  TableColumns columns() const;
-  QueryData generate(QueryContext& request);
-};
-} // namespace osquery
